Implement EmptyState handling in TransactionList and introduce Empty component
- Added an Empty component to provide user feedback when no transactions are available.
- Updated the Data component to utilize EmptyState for improved handling of empty transaction collections.
- Refactored the rendering logic to enhance user experience and maintainability.
