+++ apps/app/src/app/@buyer/transaction/ui/TransactionList/Data.tsx
@@ -1,13 +1,9 @@
-import { useLocale } from "@use-pico/client/hook";
-import { ChevronRightIcon, MessageIcon } from "@use-pico/client/icon";
 import type { MarkSuspense } from "@use-pico/client/type";
 import { Container } from "@use-pico/client/ui/container";
-import { LinkTo } from "@use-pico/client/ui/link-to";
-import { Status } from "@use-pico/client/ui/status";
-import { Tx } from "@use-pico/client/ui/tx";
-import { translator } from "@use-pico/common/translator";
+import { EmptyState } from "@use-pico/client/ui/empty-state";
 import { withTransactionQuery } from "@zbav-se.me/sdk/query/buyer/transaction";
-import type { FC, ReactNode } from "react";
+import { type FC, type ReactNode, useMemo } from "react";
+import { Empty } from "./Empty";
 
 export namespace Data {
 	export interface Props extends Container.Props, MarkSuspense.Props {
@@ -23,8 +19,7 @@ export const Data: FC<Data.Props> = ({
 	ui,
 	...props
 }) => {
-	const locale = useLocale();
-	const { data } = withTransactionQuery.useCollectionQuery(
+	const { data: transactionCollection } = withTransactionQuery.useCollectionQuery(
 		{
 			sort: [
 				{
@@ -41,86 +36,36 @@ export const Data: FC<Data.Props> = ({
 			refetchInterval,
 		},
 	);
-	const { data: transactionCount } = withTransactionQuery.useCountQuery({});
+
+	const check = useMemo(() => {
+		return [
+			{
+				check() {
+					return !transactionCollection.length;
+				},
+				render() {
+					return <Empty />;
+				},
+			},
+		] satisfies EmptyState.Check[];
+	}, [
+		transactionCollection,
+	]);
 
 	return (
 		<Container
 			ui={{
 				scroll: "vertical",
 				height: "full",
+				layout: "vertical-flex",
+				gap: "default",
 				...ui,
 			}}
 			{...props}
 		>
-			{transactionCount.isEmpty ? (
-				<Container
-					ui={{
-						layout: "vertical-centered",
-						height: "full",
-					}}
-				>
-					<Status
-						icon={MessageIcon}
-						textTitle={translator.text("No transactions as buyer (title)")}
-						textMessage={translator.text("No transactions as buyer (message)")}
-						action={
-							<LinkTo
-								icon={ChevronRightIcon}
-								iconPosition={"right"}
-								to="/$locale/buyer/feed/default"
-								params={{
-									locale,
-								}}
-								ui={{
-									background: "default",
-									border: true,
-									shadow: true,
-									round: "default",
-									size: "default",
-								}}
-							>
-								<Tx label="Go to my feed (button)" />
-							</LinkTo>
-						}
-						ui={{
-							tone: "brand",
-							theme: "light",
-							inner: "4xl",
-						}}
-						className={"text-center"}
-					/>
-				</Container>
-			) : transactionCount.isFilterEmpty ? (
-				<Container
-					ui={{
-						layout: "vertical-centered",
-						height: "full",
-					}}
-				>
-					<Status
-						icon={MessageIcon}
-						textTitle={translator.text("No transactions for current filter (title)")}
-						textMessage={translator.text(
-							"No transactions for current filter (message)",
-						)}
-						ui={{
-							tone: "brand",
-							theme: "light",
-							inner: "4xl",
-						}}
-						className={"text-center"}
-					/>
-				</Container>
-			) : (
-				<Container
-					ui={{
-						layout: "vertical-flex",
-						gap: "default",
-					}}
-				>
-					{data.map((transactionId) => renderItem(transactionId))}
-				</Container>
-			)}
+			<EmptyState check={check}>
+				{transactionCollection.map((transactionId) => renderItem(transactionId))}
+			</EmptyState>
 		</Container>
 	);
 };
